+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changes
 
+## v0.3.1
+* Fix `unwrap_html` when no result was found.
+
+## v0.3.0
+
+* Code quality improvements: linting, typing
+* Supporting Python versions 3.10 - 3.12
+* Improve Outlook forward detection
+
 ## v0.2.4
 
 * First version using auto-release process